Nothing done, just do chit

+++ src/org/academiadecodigo/PopStarsSpaceInvaders/menu/Clickable.java
@@ -1,6 +1,6 @@
 package org.academiadecodigo.PopStarsSpaceInvaders.menu;
 
 public interface Clickable {
-    void setX(int x);
-    void setY(int y);
+    void setX(double x);
+    void setY(double y);
 }
